rewrote code to search for the particular link based on today's date
