+++ src/index.ts
@@ -1,25 +1,40 @@
 import * as cheerio from "cheerio";
 import axios from "axios";
-import { day, dayToChoose, dayToReturn, getMonthName, month } from "./app";
+import { day, dayToReturn, } from "./middlewares/today";
 import { findLinkWithWord } from "./middlewares/findLink";
-
 
 const url: string = `https://www.thedevotionals.com.ng/devotional/rhapsody-of-realities/`;
 
-
-
 async function scrapper() {
   const urltoscrape = await crawler(url);
-  console.log(urltoscrape)
+  const response = await axios.get(urltoscrape);
+  const html = await response.data;
+  const $ = cheerio.load(html);
+  const d = [];
+
+  $("p").each((i, li) => {
+    d.push($(li).text());
+  });
+  // console.log(urltoscrape)
+  console.log(d);
 }
 
 async function crawler(url) {
-  const response = await axios.get(url)
-  const html = await response.data
-  const $ = cheerio.load(html);
-  const links = $('a').map((i, link) => link.attribs.href).get();
-  const result = findLinkWithWord(links, "rhapsody")
-  return result + `-for-${dayToChoose}-${dayToReturn(day)}-${getMonthName(month)}-2023`;
+  try {
+    const response = await axios.get(url);
+    const html = await response.data;
+    const $ = cheerio.load(html);
+    const links = $("a")
+      .map((i, link) => link.attribs.href)
+      .get()
+      .slice(-82);
+    const result = findLinkWithWord(links, `${dayToReturn(day)}`);
+    console.log(result);
+    return result;
+  } catch (error) {
+    console.log(error)
+  }
 }
 
-scrapper()
+scrapper();
+// crawler(url)
